Refactoring metric emission to only take testcase id as argument
